Decouple VPN code from module extensions
